feat: keep wallet connection
